[android-builder] check if sdk is supported
