+++ src/builders/ios/index.ts
@@ -1,8 +1,5 @@
-import path from 'path';
-
 import fs from 'fs-extra';
 import pick from 'lodash/pick';
-import { ExponentTools } from 'xdl';
 
 import buildArchive from 'turtle/builders/ios/archive';
 import { createBuilderContext, IContext } from 'turtle/builders/ios/context';
@@ -10,13 +7,12 @@ import buildSimulator from 'turtle/builders/ios/simulator';
 import { logErrorOnce } from 'turtle/builders/utils/common';
 import prepareAdHocBuildCredentials from 'turtle/builders/utils/ios/adhocBuild';
 import { uploadBuildToS3 } from 'turtle/builders/utils/uploader';
+import { ensureCanBuildSdkVersion } from 'turtle/builders/utils/version';
 import config from 'turtle/config';
-import { IOS, PLATFORMS } from 'turtle/constants/index';
+import { IOS } from 'turtle/constants/index';
 import { IJob, IJobResult } from 'turtle/job';
 
 const { BUILD_TYPES } = IOS;
-
-let SUPPORTED_SDK_VERSIONS: number[];
 
 export default async function iosBuilder(job: IJob): Promise<IJobResult> {
   if (job.config.buildType !== BUILD_TYPES.CLIENT) {
@@ -68,25 +64,3 @@ async function cleanup(ctx: IContext) {
 
 const getTemporaryDirs = (ctx: IContext) =>
   Object.values(pick(ctx, ['appDir', 'provisioningProfileDir']));
-
-async function ensureCanBuildSdkVersion(job: IJob) {
-  if (config.builder.useLocalWorkingDir) {
-    return;
-  }
-
-  if (!SUPPORTED_SDK_VERSIONS) {
-    SUPPORTED_SDK_VERSIONS = await findSupportedSdkVersions();
-  }
-
-  const targetMajorSdkVersion = ExponentTools.parseSdkMajorVersion(job.sdkVersion || job.manifest.sdkVersion);
-  if (!SUPPORTED_SDK_VERSIONS.includes(targetMajorSdkVersion)) {
-    throw new Error(`Unsupported SDK Version!`);
-  }
-}
-
-async function findSupportedSdkVersions(): Promise<number[]> {
-  const SDK_DIR_PREFIX = 'sdk';
-  const files = await fs.readdir(path.join(config.directories.workingDir, PLATFORMS.IOS));
-  const sdks = files.filter((file) => file.startsWith(SDK_DIR_PREFIX));
-  return sdks.map((sdk) => parseInt(sdk.substr(SDK_DIR_PREFIX.length), 10));
-}
